requirements.txt
beautifulsoup4==4.6.3
msedge-selenium-tools==3.141.3
selenium==3.141.0
